+++ src/main/java/com/microsoft/store/partnercenter/models/ResourceAttributes.java
@@ -22,7 +22,6 @@ public class ResourceAttributes
         {
             this.setObjectType( type.getSimpleName() );
         }
-
     }
 
     /**