+++ src/main/java/com/microsoft/store/partnercenter/models/ratecards/AzureOfferTerm.java
@@ -74,5 +74,4 @@ public class AzureOfferTerm extends ResourceBase
     {
     	__EffectiveDate = value;
     }
-
 }