+++ ChangeLog.md
@@ -24,6 +24,8 @@
 
 * Error Handling
   * Addressed issue with the errors not being deserialized correctly
+* Invoices
+  * Addressed issue [#47](https://github.com/microsoft/Partner-Center-Java/issues/47) where the billing provider was incorrect
 
 ## 1.13.0